It all starts here.
So, the kids have really wanted to play Octogeddon on their Android Tablets ever since it first came out a couple years ago. Since we still haven't heard any movement on an official release, we decided to make our own game that is in the same style. This is the beginning of it.